Migrate GD points routes to superfluid-points API
